requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==41.0.7
Django==5.0.4
django-cors-headers==4.6.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.4.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0


deepface==0.0.80
insightface==0.7.3
matplotlib==3.8.2
numpy==1.23.5
opencv_python==4.6.0.66
pandas==2.3.0
psycopg2==2.9.9
retina_face==0.0.13
scikit_learn==1.4.1.post1
tensorflow==2.10.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
ultralytics==8.3.68

